Create tree-decrements.py
This is similar to a practice question I am trying to solve. You have been given an undirected tree with an array of values. A value val[i] is associated with the ith node. In a single operation, two nodes can be selected, and their values can be decremented by 1 at a cost equal to the distance between the two nodes, i.e., the number of edges in the simple path between them. It is possible to select the same node for the operation and decrease its value by 2 at the cost of 0.

t_from = [1, 1, 3, 5], t_to = [2, 3, 4, 5] val = [3, 2, 4, 2, 5]



Steps:
1. convert all val to %2 (1 or 0)
2. adjacent list for each val (neighbours)
3. find if each val is a leaf or parent
4. iterate the leaves and find their parents
5. check if the parent has other child (if not, it becomes leaf)
6. at the end, remaining with 2 nodes left. check if the root is 1 or 0. if it is 1 then add to the totalcost.